ci: run the full check once per pull request
The full lint, type, test, and build run twice on every pull request: once on
the Linux runner and again inside verify.sh on the macOS one, which reports
nothing the first run did not and costs macOS minutes to say it. CI's macOS
job now captures the fixture evidence alone, and the bundle assertions
test-macos.sh made move into evidence.sh, where the frames are already cut
from a packaged app — so the one script that packages is also the one that
checks what packaging carried, and test-macos.sh goes with its duplicate
check.

The release workflow loses its own duplicates the same way: the publish step
is scripts/release/publish-github.sh, which a by-hand release already runs,
the notarization step stops repeating the signature verification the step
above it just made, and five node invocations that each loaded the same
release-config module become one. The pnpm, Node, and bootstrap triple every
job restated is the composite action in .github/actions/setup.

The workspace Node floor was still 22.18 while .nvmrc, every manifest, and
the engine-drift check in repository-checks.sh had moved to 24.

The deleted tests are the ones that read a shell script or a workflow with a
regular expression and assert on the text: they pin wording rather than
behavior, and they fail on the edits above without a release being any less
verified. The iOS project tests keep their structure and privacy-manifest
assertions and move beside the project they read; the nine that matched
regular expressions against Swift and scheme sources go.
