+++ .github/actions/setup/action.yml
@@ -1,0 +1,13 @@
+name: Set up the Luke workspace
+description: Pin pnpm and Node from .nvmrc, then install the locked dependencies.
+
+runs:
+  using: composite
+  steps:
+    - uses: pnpm/action-setup@b906affcce14559ad1aafd4ab0e942779e9f58b1 # v4
+    - uses: actions/setup-node@249970729cb0ef3589644e2896645e5dc5ba9c38 # v6
+      with:
+        node-version-file: .nvmrc
+        cache: pnpm
+    - run: ./scripts/bootstrap.sh
+      shell: bash
