+++ scripts/evidence.sh
@@ -8,6 +8,7 @@ source "$SCRIPT_DIRECTORY/lib/workspace.sh"
 sidecar_require_macos
 sidecar_require_node
 sidecar_require_command sips
+sidecar_require_command plutil
 sidecar_ensure_dependencies
 
 cd "$SIDECAR_REPO_ROOT"
@@ -15,7 +16,30 @@ pnpm package
 PACKAGED_APP=$(node -e "import('./apps/desktop/scripts/package-layout.mjs').then((layout) => process.stdout.write(layout.packagedAppPath(process.cwd())))")
 APP_EXECUTABLE="$PACKAGED_APP/Contents/MacOS/Luke"
 if [[ ! -x "$APP_EXECUTABLE" ]]; then
-    printf 'error: packaged app executable was not found\n' >&2
+    printf 'error: electron-builder did not produce a runnable app at %s\n' "$PACKAGED_APP" >&2
+    exit 1
+fi
+# The captures below launch the packaged app, so they prove the bundle runs but
+# not that electron-builder carried everything into it: a helper, the addon, or
+# the icon can be missing from a bundle whose panel still paints.
+for helper in mac-screen-geometry mac-talk-key; do
+    if [[ ! -x "$PACKAGED_APP/Contents/Resources/$helper" ]]; then
+        printf 'error: packaged app is missing the %s helper\n' "$helper" >&2
+        exit 1
+    fi
+done
+if [[ ! -f "$PACKAGED_APP/Contents/Resources/mac-stationary-window.node" ]]; then
+    printf 'error: packaged app is missing the mac-stationary-window.node addon\n' >&2
+    exit 1
+fi
+BUNDLE_ICON_FILE=$(plutil -extract CFBundleIconFile raw -o - "$PACKAGED_APP/Contents/Info.plist")
+PACKAGED_ICON="$PACKAGED_APP/Contents/Resources/$BUNDLE_ICON_FILE"
+if [[ ! -f "$PACKAGED_ICON" ]]; then
+    printf 'error: packaged app is missing its declared icon: %s\n' "$PACKAGED_ICON" >&2
+    exit 1
+fi
+if ! cmp -s "$PACKAGED_ICON" "$SIDECAR_DESKTOP_APP_ROOT/.build/Luke.icns"; then
+    printf 'error: packaged app icon does not match the generated Luke icon\n' >&2
     exit 1
 fi
